agent/requirements.txt
bedrock-agentcore
strands-agents
fastapi
boto3
pyyaml
